Updated: replace Navigator.pop() with Get.back().
