+++ lib/views/home/home_view.dart
@@ -1,16 +1,13 @@
-import 'package:test_task_app/controllers/task/create_task_controller.dart';
-
 import '../../constants/constants.dart';
 import '../common_widgets/create_app_bar_widget.dart';
 import '../task/task_create_view.dart';
-import 'widgets/task_list.dart';
+import 'widgets/task_list_view.dart';
 
 class HomeView extends StatelessWidget {
   const HomeView({super.key});
 
   @override
   Widget build(BuildContext context) {
-    TaskController taskController = Get.put(TaskController());
     return Scaffold(
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       floatingActionButton: FloatingActionButton(
@@ -20,7 +17,7 @@ class HomeView extends StatelessWidget {
         },
       ),
       appBar: createAppBarWidget(title: homeTitle, implyLeading: false),
-      body: const SafeArea(child: TaskList()),
+      body: const SafeArea(child: TaskListView()),
     );
   }
 }
